Move per-site configuration into a separate task
According to my old comments, this did not work with the version of
Ansible I was using.

This is better organization, and it drastically simplifies the step that
my old comment described as "clear as mud".

In addition to refactoring, I updated syntax for the current version of
Ansible and may have made some other fixes.